core: Rename FileStreamProvider to Reader
- Shorter
- Better match Writer
